+++ organize/actions/shell.py
@@ -43,9 +43,10 @@ class Shell(Action):
 
     def pipeline(self, args: dict, simulate: bool):
         full_cmd = self.cmd.render(**args)
-        self.print("$ %s" % full_cmd)
+
         if not simulate or self.run_in_simulation:
             # we use call instead of run to be compatible with python < 3.5
+            self.print("$ %s" % full_cmd)
             logger.info('Executing command "%s" in shell.', full_cmd)
             try:
                 call = subprocess.run(
@@ -70,6 +71,14 @@ class Shell(Action):
                         "returncode": e.returncode,
                     }
                 }
+        else:
+            self.print("** not run in simulation ** $ %s" % full_cmd)
+            return {
+                self.get_name(): {
+                    "output": "**simulation**",
+                    "returncode": 0,
+                }
+            }
 
     def __str__(self) -> str:
         return 'Shell(cmd="%s")' % self.cmd